[Refactor] Introduce platform module
