+++ be/src/platform/download_util.cpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "util/download_util.h"
+#include "platform/download_util.h"
 
 #include "base/uuid/uuid_generator.h"
 #include "fmt/format.h"
 #include "fs/fs_options.h"
-#include "udf/udf_downloader.h"
+#include "platform/udf_downloader.h"
 
 namespace starrocks {
