+++ be/src/platform/download_util.h
