+++ be/src/runtime/jdbc_driver_manager.cpp
@@ -26,7 +26,7 @@
 #include "fs/fs.h"
 #include "fs/fs_util.h"
 #include "gutil/strings/split.h"
-#include "util/download_util.h"
+#include "platform/download_util.h"
 
 namespace starrocks {
