+++ be/src/runtime/user_function_cache.cpp
@@ -49,7 +49,7 @@
 #include "fs/fs.h"
 #include "fs/fs_util.h"
 #include "gutil/strings/split.h"
-#include "util/download_util.h"
+#include "platform/download_util.h"
 
 namespace starrocks {
 static const int kLibShardNum = 128;
